BO6: Moved generator components into their own folder
