+++ components/black-ops-six/BlackOpsSixZombiesLoadout.tsx
@@ -2,14 +2,14 @@ import { useEffect, useState } from "react";
 import { Container, Row, Col } from "react-bootstrap";
 import Button from "react-bootstrap/Button";
 //Helpers
-import { implodeObject } from "../helpers/implodeObject";
-import { fetchWeapon } from "../helpers/fetchWeapon";
+import { implodeObject } from "../../helpers/implodeObject";
+import { fetchWeapon } from "../../helpers/fetchWeapon";
 import { fetchAttachments } from "@/helpers/fetchAttachments";
 import { fetchEquipment } from "@/helpers/fetchEquipment";
 import { fetchBO6Gobblegums } from "@/helpers/generator/black-ops-six/fetchBO6Gobblegums";
 import { fetchBO6ZombiesMap } from "@/helpers/generator/black-ops-six/fetchBO6ZombiesMap";
 //Styles
-import "../public/styles/components/Loadout.css";
+import "@/public/styles/components/Loadout.css";
 
 function BlackOpsSixZombiesLoadout() {
   const [containerClass, setContainerClass] = useState("hidden");
